+++ CHANGELOG.md
@@ -8,6 +8,26 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
+## [3.2.7] - 2025-09-??
+
+### Breaking Changes
+
+- [platform]: Typed PlatformConfig with the default properties: name, type, version, debug and unregisterOnShutdown.
+- [platform]: Removed DefaultPlatformConfig interface definition.
+
+### Added
+
+### Changed
+
+- [package]: Updated dependencies.
+- [matterbridge.io]: Updated site.
+
+### Fixed
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="bmc-button.svg" alt="Buy me a coffee" width="80">
+</a>
+
 ## [3.2.6] - 2025-09-06
 
 ### Added